+++ rust-crates/sns-sdk/src/blocking/mod.rs
@@ -5,3 +5,5 @@ pub mod record_v1;
 pub mod record_v2;
 pub mod resolve;
 pub mod subdomain;
+
+mod rpc;