+++ rust-crates/sns-sdk/src/domain.rs
@@ -1,0 +1,7 @@
+use solana_program::pubkey::Pubkey;
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct SolDomain {
+    pub domain: String,
+    pub key: Pubkey,
+}