+++ rust-crates/sns-sdk/src/nft.rs
@@ -6,3 +6,10 @@ pub struct SnsNftDomain {
     pub key: Pubkey,
     pub mint: Pubkey,
 }
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct SolNftDomain {
+    pub domain: String,
+    pub key: Pubkey,
+    pub mint: Pubkey,
+}